+++ src/App.tsx
@@ -3,10 +3,10 @@ import { ChakraProvider } from '@chakra-ui/react'
 import RootRouter from './route/RootRouter'
 import GlobalStyle from './style/Global'
 import { useDispatch } from 'react-redux'
-import { setLogin, setLogout } from './redux/auth/authSlice'
-import { getUsersMe } from './utils/api'
-import theme from './style/theme'
-import Footer from './components/Footer'
+import { setLogin, setLogout } from '@redux/auth/authSlice'
+import { getUsersMe } from '@api/userAPI'
+import theme from '@style/theme'
+import Footer from '@components/Footer'
 import { useEffect } from 'react'
 
 const App = (): JSX.Element => {
