+++ src/components/NavBar/MyMenu/index.tsx
@@ -1,7 +1,7 @@
 import { Menu, MenuButton, MenuItem, MenuList } from '@chakra-ui/react'
-import { setLogout } from '../../../redux/auth/authSlice'
-import { useAppSelect } from '../../../redux/store.hooks'
-import { postLogout } from '../../../utils/api'
+import { setLogout } from '@redux/auth/authSlice'
+import { useAppSelect } from '@redux/store.hooks'
+import { postLogout } from '@api/userAPI'
 import { DesktopMenu, MobileMenu } from './styles'
 import { ItemName } from '../styles'
 import { useDispatch } from 'react-redux'
