+++ src/components/NavBar/MyMenu/styles.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import theme from '../../../style/theme'
+import theme from '@style/theme'
 
 const DesktopMenu = styled.div`
   display: none;
