+++ src/pages/AnsweredPetitions/index.tsx
@@ -3,7 +3,7 @@ import { Stack } from '@chakra-ui/react'
 import { PetitionsText, PetitionsTitle } from './styles'
 import PaginationButtons from '../../components/PaginationButtons'
 import PetitionList from '../../components/PetitionList'
-import { getAnsweredPetitionsByQuery } from '../../utils/api/'
+import { getAnsweredPetitionsByQuery } from '@api/petitionAPI'
 import { useNavigate } from 'react-router-dom'
 import Inner from '../../components/Inner'
