+++ src/pages/Home/styles.ts
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled'
-import theme from '../../style/theme'
-import MainImg from '../../assets/img/gist_summer.jpg'
+import theme from '@style/theme'
+import MainImg from '@assets/img/gist_summer.jpg'
 import { keyframes } from '@emotion/react'
 import { Box } from '@chakra-ui/react'
