+++ src/route/RootRouter.tsx
@@ -9,7 +9,6 @@ import Write from '../pages/Write'
 import MyPetitions from '../pages/MyPetitions'
 import AnsweredPetitions from '../pages/AnsweredPetitions'
 import { AuthRoute } from './PrivateRouter'
-// import { AuthRoute, UnauthRoute } from './PrivateRouter'
 import FindingPassword from '../pages/FindingPassword'
 import NavBar from '../components/NavBar'
 
@@ -19,12 +18,8 @@ const RootRouter = (): JSX.Element => {
       <NavBar></NavBar>
       <Routes>
         <Route path="/" element={<Home />} />
-        {/* <Route path="/login" element={<UnauthRoute />}> */}
         <Route path="/login" element={<Login />} />
-        {/* </Route> */}
-        {/* <Route path="/register" element={<UnauthRoute />}> */}
         <Route path="/register" element={<Register />} />
-        {/* </Route> */}
         <Route path="/write" element={<AuthRoute />}>
           <Route index element={<Write />} />
         </Route>
