+++ tsconfig.extend.json
@@ -8,7 +8,8 @@
       "@style/*": ["src/style/*"],
       "@components/*": ["src/components/*"],
       "@api/*": ["src/utils/api/*"],
-      "@redux/*": ["src/redux/*"]
+      "@redux/*": ["src/redux/*"],
+      "@utils/*": ["src/utils/*"]
     }
   }
 }
